chore(automated testing): updated to use Mocha
migrates mjml-core and mjml-parser-xml from Node to Mocha, aligning with the existing Mocha-based testing style in the repo.
