+++ packages/mjml-core/package.json
@@ -19,7 +19,7 @@
   "scripts": {
     "clean": "rimraf lib",
     "build": "babel src --out-dir lib --root-mode upward",
-    "test": "node ./tests/index.js"
+    "test": "mocha ./tests/*.test.js"
   },
   "dependencies": {
     "@babel/runtime": "^7.28.4",
