--- packages/mjml-core/tests/index.js
@@ -1,6 +1,0 @@
-require('./jsonToXml-test')
-require('./mergeOutlookConditionnals-test')
-require('./minifyOutlookConditionnals-test')
-require('./shorthandParser-test')
-require('./skeleton-test')
-require('./widthParser-test')
