+++ packages/mjml-core/tests/shorthandParser.test.js
@@ -1,0 +1,44 @@
+const chai = require('chai')
+const helper = require('../lib/helpers/shorthandParser')
+
+const shorthandParser = helper && helper.default
+
+describe('shorthandParser', () => {
+  const directions = ['top', 'right', 'bottom', 'left']
+
+  const testValues = [
+    {
+      input: '1px',
+      output: { top: 1, right: 1, bottom: 1, left: 1 },
+    },
+    {
+      input: '1px 0',
+      output: { top: 1, right: 0, bottom: 1, left: 0 },
+    },
+    {
+      input: '1px 2px 3px',
+      output: { top: 1, right: 2, bottom: 3, left: 2 },
+    },
+    {
+      input: '1px 2px 3px 4px',
+      output: { top: 1, right: 2, bottom: 3, left: 4 },
+    },
+    {
+      input: ' 1px 2px  3px 4px ',
+      output: { top: 1, right: 2, bottom: 3, left: 4 },
+    },
+  ]
+
+  testValues.forEach(({ input, output }, index) => {
+    it(`parses shorthand values correctly (case ${index + 1})`, () => {
+      directions.forEach((dir) => {
+        chai
+          .expect(
+            shorthandParser(input, dir),
+            `shorthandParser test failed for ${dir}`,
+          )
+          .to.deep.equal(output[dir])
+      })
+    })
+  })
+})
