+++ packages/mjml-parser-xml/package.json
@@ -19,7 +19,7 @@
   "scripts": {
     "clean": "rimraf lib",
     "build": "babel src --out-dir lib --root-mode upward",
-    "test": "node ./test/test.js"
+    "test": "mocha ./test/*.test.js"
   },
   "dependencies": {
     "@babel/runtime": "^7.28.4",
